@@ -3,101 +3,42 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "build": "vue-cli-service build",
-    "test": "jest",
-    "lint": "vue-cli-service lint",
-    "dev": "concurrently \"vue-cli-service serve --open\" \"nodemon server/index.js\"",
-    "heroku-postbuild": "npm install --only=dev --no-shrinkwrap && npm run build",
-    "start": "node server/index.js",
-    "test:unit": "vue-cli-service test:unit"
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview",
+    "test": "vitest",
+    "lint": "eslint . --ext .vue,.js,.jsx,.cjs,.mjs --fix --ignore-path .gitignore"
   },
   "dependencies": {
-    "babel-polyfill": "^6.26.0",
     "compression": "^1.7.4",
-    "dns": "^0.2.2",
-    "dotenv": "^5.0.1",
-    "express": "^4.16.4",
-    "google-maps": "^3.3.0",
-    "gsap": "^1.20.6",
-    "intersection-observer": "^0.5.1",
-    "nodemailer": "^4.7.0",
-    "nodemon": "^1.18.10",
+    "dotenv": "^16.4.5",
+    "express": "^4.18.3",
+    "google-maps": "^4.3.3",
+    "gsap": "^3.12.5",
+    "intersection-observer": "^0.12.0",
+    "nodemailer": "^6.9.12",
+    "nodemon": "^3.1.0",
     "path": "^0.12.7",
-    "register-service-worker": "^1.6.2",
-    "vue": "^2.6.10",
-    "vue-loader": "^15.7.0",
-    "vue-scrollto": "^2.15.0",
-    "vue-tiny-lazyload-img": "^0.1.0",
-    "whatwg-fetch": "^3.0.0"
+    "pinia": "^2.1.7",
+    "register-service-worker": "^1.7.2",
+    "vue": "^3.4.21",
+    "vue-router": "^4.3.0",
+    "vue-scrollto": "^2.20.0",
+    "whatwg-fetch": "^3.6.20"
   },
   "devDependencies": {
-    "@babel/core": "^7.4.3",
-    "@babel/preset-env": "^7.4.3",
-    "@vue/cli-plugin-babel": "^3.5.5",
-    "@vue/cli-plugin-eslint": "^3.5.1",
-    "@vue/cli-plugin-pwa": "^3.5.1",
-    "@vue/cli-plugin-unit-jest": "^3.5.3",
-    "@vue/cli-service": "^3.5.3",
-    "@vue/eslint-config-prettier": "^4.0.1",
-    "@vue/test-utils": "^1.0.0-beta.20",
-    "babel-eslint": "^10.0.1",
-    "babel-jest": "^24.7.1",
-    "babel-preset-env": "^1.7.0",
-    "concurrently": "^3.6.1",
-    "eslint": "^5.16.0",
-    "eslint-plugin-vue": "^5.0.0",
-    "jest": "^24.7.1",
-    "node-sass": "^4.11.0",
-    "prettier": "^1.16.4",
-    "sass-loader": "^6.0.7",
-    "style-loader": "^0.20.3",
-    "vue-jest": "^3.0.4",
-    "vue-template-compiler": "^2.6.10"
-  },
-  "babel": {
-    "env": {
-      "test": {
-        "presets": [
-          [
-            "env",
-            {
-              "targets": {
-                "node": "current"
-              }
-            }
-          ]
-        ]
-      }
-    }
-  },
-  "postcss": {
-    "plugins": {
-      "autoprefixer": {}
-    }
-  },
-  "browserslist": [
-    "> 1%",
-    "last 2 versions",
-    "not ie <= 8"
-  ],
-  "jest": {
-    "moduleFileExtensions": [
-      "js",
-      "vue"
-    ],
-    "moduleNameMapper": {
-      "^@/(.*)$": "<rootDir>/resources/assets/js/components/$1"
-    },
-    "transform": {
-      "^.+\\.js$": "<rootDir>/node_modules/babel-jest",
-      ".*\\.(vue)$": "<rootDir>/node_modules/vue-jest"
-    },
-    "snapshotSerializers": [
-      "<rootDir>/node_modules/jest-serializer-vue"
-    ]
+    "@vitejs/plugin-vue": "^5.0.4",
+    "@vue/eslint-config-prettier": "^9.0.0",
+    "@vue/test-utils": "^2.4.4",
+    "eslint": "^8.57.0",
+    "eslint-plugin-vue": "^9.23.0",
+    "prettier": "^3.2.5",
+    "sass": "^1.72.0",
+    "vite": "^5.1.4",
+    "vitest": "^1.3.1"
   },
   "engines": {
-    "node": "v8.8.1",
-    "npm": "6.1.0"
+    "node": ">=18.0.0",
+    "npm": ">=8.0.0"
   }
 }
