@@ -39,11 +39,7 @@
             >
           </li>
           <li class="nav-list-item">
-            <a
-              class="nav-item"
-              :href="resumeLink"
-              download="LeonardoVelazquezResume"
-            >
+            <a class="nav-item" :href="resumeLink" download="LeonardoVelazquezResume">
               <i class="icon-download-alt"></i>
               Resume
             </a>
@@ -63,56 +59,57 @@
 </template>
 
 <script>
-import NavLink from "./NavLink";
-import BrandLogo from "./BrandLogo";
-import EventBus from "../EventBus";
+import NavLink from './NavLink.vue'
+import BrandLogo from './BrandLogo.vue'
+import EventBus from '../EventBus'
 
 export default {
-  name: "MainNav",
+  name: 'MainNav',
   components: {
     BrandLogo,
-    "nav-link": NavLink
+    'nav-link': NavLink,
   },
   props: {
     resumeLink: String,
     currentSectionId: String,
-    navOffset: Number
+    navOffset: Number,
   },
   data() {
     return {
       isMenuOpen: false,
-      isNavHidden: true
-    };
+      isNavHidden: true,
+    }
   },
   mounted() {
-    EventBus.$on("routechange", this.openMenu);
-    window.addEventListener("scroll", this.updateNav);
+    EventBus.on('routechange', this.openMenu)
+    window.addEventListener('scroll', this.updateNav)
   },
-  destroyed() {
-    window.removeEventListener("scroll", this.updateNav);
+  unmounted() {
+    window.removeEventListener('scroll', this.updateNav)
+    EventBus.off('routechange', this.openMenu)
   },
   updated() {},
   methods: {
     updateNav() {
-      this.isNavHidden = window.scrollY === 0;
+      this.isNavHidden = window.scrollY === 0
     },
     openMenu() {
-      this.isMenuOpen = false;
-    }
-  }
-};
+      this.isMenuOpen = false
+    },
+  },
+}
 </script>
 
 <!-- Add "scoped" attribute to limit CSS to this component only -->
 <style lang="scss" scoped>
-@import "../styles/settings";
-@import "../styles/media-queries";
+@use '../styles/settings.scss' as *;
+@use '../styles/media-queries.scss' as media;
 
 .navbar {
   position: fixed;
   top: 0;
   width: 100%;
-  font-family: Montserrat, "Helvetica Neue", Helvetica, Arial, sans-serif;
+  font-family: Montserrat, 'Helvetica Neue', Helvetica, Arial, sans-serif;
   margin-bottom: 0;
   text-transform: uppercase;
   z-index: 1000;
@@ -151,10 +148,10 @@ export default {
       color: $white-color;
       background-color: $black-color;
     }
-    @media (min-width: $bp-ms) {
+    @media (min-width: media.$bp-ms) {
       padding: 1.7rem 1rem 1.7rem;
     }
-    @media (min-width: $bp-lg) {
+    @media (min-width: media.$bp-lg) {
       padding: 2rem;
     }
   }
@@ -168,7 +165,7 @@ export default {
   display: block;
   padding: 0;
   margin: 0;
-  @media (min-width: $bp-ms) {
+  @media (min-width: media.$bp-ms) {
     display: flex;
   }
 }
@@ -192,7 +189,7 @@ export default {
 
 li:hover .download-icon {
   background-size: cover;
-  background-image: url(data:image/svg+xml;utf8;base64,PD94bWwgdmVyc2lvbj0iMS4wIiBlbmNvZGluZz0iaXNvLTg4NTktMSI/Pgo8IS0tIEdlbmVyYXRvcjogQWRvYmUgSWxsdXN0cmF0b3IgMTYuMC4wLCBTVkcgRXhwb3J0IFBsdWctSW4gLiBTVkcgVmVyc2lvbjogNi4wMCBCdWlsZCAwKSAgLS0+CjwhRE9DVFlQRSBzdmcgUFVCTElDICItLy9XM0MvL0RURCBTVkcgMS4xLy9FTiIgImh0dHA6Ly93d3cudzMub3JnL0dyYXBoaWNzL1NWRy8xLjEvRFREL3N2ZzExLmR0ZCI+CjxzdmcgeG1sbnM9Imh0dHA6Ly93d3cudzMub3JnLzIwMDAvc3ZnIiB4bWxuczp4bGluaz0iaHR0cDovL3d3dy53My5vcmcvMTk5OS94bGluayIgdmVyc2lvbj0iMS4xIiBpZD0iQ2FwYV8xIiB4PSIwcHgiIHk9IjBweCIgd2lkdGg9IjUxMnB4IiBoZWlnaHQ9IjUxMnB4IiB2aWV3Qm94PSIwIDAgNDMzLjUgNDMzLjUiIHN0eWxlPSJlbmFibGUtYmFja2dyb3VuZDpuZXcgMCAwIDQzMy41IDQzMy41OyIgeG1sOnNwYWNlPSJwcmVzZXJ2ZSI+CjxnPgoJPGcgaWQ9ImZpbGUtZG93bmxvYWQiPgoJCTxwYXRoIGQ9Ik0zOTUuMjUsMTUzaC0xMDJWMGgtMTUzdjE1M2gtMTAybDE3OC41LDE3OC41TDM5NS4yNSwxNTN6IE0zOC4yNSwzODIuNXY1MWgzNTd2LTUxSDM4LjI1eiIgZmlsbD0iI0ZGRkZGRiIvPgoJPC9nPgo8L2c+CjxnPgo8L2c+CjxnPgo8L2c+CjxnPgo8L2c+CjxnPgo8L2c+CjxnPgo8L2c+CjxnPgo8L2c+CjxnPgo8L2c+CjxnPgo8L2c+CjxnPgo8L2c+CjxnPgo8L2c+CjxnPgo8L2c+CjxnPgo8L2c+CjxnPgo8L2c+CjxnPgo8L2c+CjxnPgo8L2c+Cjwvc3ZnPgo=);
+  background-image: url(data:image/svg+xml;utf8;base64,PD94bWwgdmVyc2lvbj0iMS4wIiBlbmNvZGluZz0iaXNvLTg4NTktMSI/Pgo8IS0tIEdlbmVyYXRvcjogQWRvYmUgSWxsdXN0cmF0b3IgMTYuMC4wLCBTVkcgRXhwb3J0IFBsdWctSW4gLiBTVkcgVmVyc2lvbjogNi4wMCBCdWlsZCAwKSAgLS0+CjwhRE9DVFlQRSBzdmcgUFVCTElDICItLy9XM0MvL0RURCBTVkcgMS4xLy9FTiIgImh0dHA6Ly93d3cudzMub3JnL0dyYXBoaWNzL1NWRy8xLjEvRFREL3N2ZzExLmR0ZCI+CjxzdmcgeG1sbnM9Imh0dHA6Ly93d3cudzMub3JnLzIwMDAvc3ZnIiB4bWxuczp4bGluaz0iaHR0cDovL3d3dy53My5vcmcvMTk5OS94bGluayIgdmVyc2lvbj0iMS4xIiBpZD0iQ2FwYV8xIiB4PSIwcHgiIHk9IjBweCIgd2lkdGg9IjUxMnB4IiBoZWlnaHQ9IjUxMnB4IiB2aWV3Qm94PSIwIDAgNDMzLjUgNDMzLjUiIHN0eWxlPSJlbmFibGUtYmFja2dyb3VuZDpuZXcgMCAwIDQzMy41IDQzMy41OyIgeG1sOnNwYWNlPSJwcmVzZXJ2ZSI+CjxnPgoJPGcgaWQ9ImZpbGUtZG93bmxvYWQiPgoJCTxwYXRoIGQ9Ik0zOTUuMjUsMTUzaC0xMDJWMGgtMTUzdjE1M2gtMTAybDE3OC41LDE3OC41TDM5NS4yNSwxNTN6IE0zOC4yNSwzODIuNXY1MWgzNTd2LTUxSDM4LjI1eiIgZmlsbD0iI0ZGRkZGRiIvPgoJPC9nPgo8L2c+CjxnPgo8L2c+CjxnPgo8L2c+CjxnPgo8L2c+CjxnPgo8L2c+CjxnPgo8L2c+CjxnPgo8L2c+CjxnPgo8L2c+CjxnPgo8L2c+CjxnPgo8L2c+Cjwvc3ZnPgo=);
 }
 
 .navbar-logo {
@@ -227,13 +224,13 @@ li:hover .download-icon {
   background-color: $white-color;
   padding: 3% 0.5rem;
   top: 0;
-  @media (min-width: $bp-sm) {
+  @media (min-width: media.$bp-sm) {
     padding: 0.75rem;
     height: 4rem;
   }
 }
 
-@media (min-width: $bp-ms) {
+@media (min-width: media.$bp-ms) {
   li {
     display: inline-block;
     font-weight: 500;
