@@ -1,8 +1,8 @@
-export { default as Background } from "./Background";
-export { default as BrandLogo } from "./BrandLogo";
-export { default as Logo } from "./Logo";
-export { default as Loader } from "./Loader";
-export { default as Map } from "./Map";
-export { default as MainNav } from "./MainNav";
-export { default as NavLink } from "./NavLink";
-export { default as Project } from "./Project";
+export { default as Background } from './Background.vue'
+export { default as BrandLogo } from './BrandLogo.vue'
+export { default as Logo } from './Logo.vue'
+export { default as Loader } from './Loader.vue'
+export { default as Map } from './Map.vue'
+export { default as MainNav } from './MainNav.vue'
+export { default as NavLink } from './NavLink.vue'
+export { default as Project } from './Project.vue'
