@@ -1,4 +1,4 @@
-export { default as About } from "./About";
-export { default as Intro } from "./Intro";
-export { default as Contact } from "./Contact";
-export { default as Work } from "./Work";
+export { default as About } from './About.vue'
+export { default as Intro } from './Intro.vue'
+export { default as Contact } from './Contact.vue'
+export { default as Work } from './Work.vue'
